+++ frontend/flutter_gestion_contable/README.md
@@ -50,3 +50,67 @@ Vamos a usar una estructura de proyecto bien organizada y separada. La idea es t
 - Archivos en `snake_case`
 - Clases en `PascalCase`
 - Variables y funciones en `camelCase`
+
+---------------------------------------------------------------------
+
+
+# Documentación para Levantar el Entorno de Flutter
+
+## Requisitos Previos
+
+- Tener instalado [Flutter SDK](https://docs.flutter.dev/get-started/install)
+- Tener instalado [Android Studio](https://developer.android.com/studio) o [Visual Studio Code](https://code.visualstudio.com/) con el plugin de Flutter
+- Tener configurado un emulador Android/iOS o un dispositivo físico
+
+---
+
+## Pasos para Configurar el Entorno
+
+### 1. Clonar el repositorio
+
+```bash
+git clone https://github.com/usuario/Proyecto-EstudioContable-.git
+cd Proyecto-EstudioContable-/frontend/flutter_gestion_contable
+```
+
+### 2. Instalar dependencias
+
+Ejecuta en la terminal dentro de la carpeta del proyecto Flutter:
+
+```bash
+flutter pub get
+```
+
+### 3. Verificar configuración
+
+Comprueba que no haya errores en el archivo `pubspec.yaml` y que todas las dependencias se hayan instalado correctamente.
+
+### 4. Ejecutar la aplicación
+
+Para correr la app en un emulador o dispositivo conectado:
+
+```bash
+flutter run
+```
+
+### 5. Compilar para web (opcional)
+
+Si quieres levantar la app en modo web:
+
+```bash
+flutter run -d chrome
+```
+
+---
+
+## Notas
+
+- Si agregas nuevas dependencias, recuerda ejecutar `flutter pub get` nuevamente.
+- Para solucionar errores de dependencias, revisa el archivo `pubspec.yaml` y asegúrate de que las versiones sean compatibles.
+- Puedes usar `flutter doctor` para verificar que tu entorno esté correctamente configurado.
+
+---
+
+## Recursos Útiles
+
+- [Documentación oficial de Flutter](https://docs.flutter.dev/)
