+++ frontend/flutter_gestion_contable/lib/main.dart
@@ -2,7 +2,7 @@
 
 import 'package:flutter/material.dart';
 import 'core/theme/app_theme.dart';
-import 'screensy/login/login_handler.dart';
+import 'screens/login/login_handler.dart';
 import 'package:http/http.dart' as http; // Import the http package
 import 'dart:convert'; // Import for JSON decoding
