+++ frontend/flutter_gestion_contable/lib/screens/main_website/main_styles.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class AppColors {
+  static const Color primary = Color(0xFF986d35);
+  static const Color secondary = Color(0xFF00a2d3);
+  static const Color background = Colors.white;
+  static const Color surfaceColor = Color(0xFFF5F5F5);
+  static const Color textPrimary = Color(0xFF212121);
+  static const Color textSecondary = Color(0xFF757575);
+  static const Color error = Color(0xFFD32F2F);
+  static const Color success = Color(0xFF388E3C);
+}
