+++ frontend/flutter_gestion_contable/lib/screens/main_website/notification_modal.dart
@@ -1,0 +1,51 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_gestion_contable/core/theme/app_colors.dart';
+
+class NotificationModal extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Dialog(
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: Container(
+        width: 300,
+        height: 500, // Ajusta el tamaño según necesites
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(8),
+          color: Colors.white,
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            // Barra superior con botón de cerrar
+            Container(
+              padding: EdgeInsets.symmetric(horizontal: 16, vertical: 10),
+              decoration: BoxDecoration(
+                color: AppColors.primary,
+                borderRadius: BorderRadius.vertical(top: Radius.circular(8)),
+              ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  Text("Notificaciones", style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold)),
+                  GestureDetector(
+                    onTap: () => Navigator.of(context).pop(),
+                    child: Icon(Icons.close, color: Colors.white),
+                  ),
+                ],
+              ),
+            ),
+
+            // Contenido del modal
+            Expanded(
+              child: Center(
+                child: Text("Aquí van las notificaciones"),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
