+++ frontend/flutter_gestion_contable/lib/screens/password_reset/password_reset_styles.dart
@@ -1,0 +1,4 @@
+//password_reset_styles.dart
+
+// Agrega la constante para el ancho máximo
+const double formMaxWidth = 400.0;
