--- frontend/flutter_gestion_contable/lib/screensy/login/login_styles.dart
@@ -1,7 +1,0 @@
-// loading_styles.dart
-
-// Tamaño de la imagen del logo
-const double logoHeight = 200.0;
-
-// Agrega la constante para el ancho máximo
-const double formMaxWidth = 600.0;
