--- frontend/flutter_gestion_contable/lib/screensy/password_reset/password_reset_styles.dart
@@ -1,4 +1,0 @@
-//password_reset_styles.dart
-
-// Agrega la constante para el ancho máximo
-const double formMaxWidth = 400.0;
