+++ frontend/flutter_gestion_contable/pubspec.yaml
@@ -11,6 +11,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  http: ^1.1.0
   provider: ^6.1.2 # Actualiza la versión aquí
   cupertino_icons: ^1.0.8
